requirements.txt
# Main Modules
django~=4.2
djangorestframework==3.15.2
psycopg[binary]

# Development modules
markdown==3.6

# Deployment modules

